requirements.txt
python-binance>=1.0.19
python-telegram-bot>=21.0
APScheduler>=3.10
google-generativeai>=0.8
aiosqlite>=0.20
ccxt>=4.0
aiohttp>=3.9
websockets>=12.0
requests>=2.31
python-dotenv>=1.0
